flac: add (unsupported) metadata encoding
This change adds enc.go, which was removed in mewkiz/flac#32.
To fix the build, a single change is made to enc.go, namely, updating
the type of bw in the encoder struct from bitio.Writer to *bitio.Writer.

Note: encoding metadata with mewkiz/flac is unsupported (for now), so
maintenance of the metadata encoding feature will be governed by the Go
community at large, when such need arrise.

Furthermore, the metadata encoding as implemented by enc.go has not been
tested locally. It is quite possible that is has broken, but if not, feel
free to use!

To use FLAC metadta encoding, pin the "unsupported-metadata-encoding"
branch of mewkiz/flac in your go.mod.

ref: https://github.com/mewkiz/flac/issues/40#issuecomment-992317606